+++ pyblog/colored_print.py
@@ -1,6 +1,7 @@
 from __future__ import print_function
 import inspect
 import sys
+import os
 
 HEADER = '\033[95m'
 OKBLUE = '\033[94m'
@@ -26,6 +27,16 @@ prefix = dict(
     , DEBUG = DEBUG
     )
 
+twordpressdir = os.path.join(os.environ['HOME'], "twordpress")
+if not os.path.isdir(twordpressdir):
+    os.makedirs(twordpressdir)
+
+debugFileName = os.path.join(twordpressdir, "log") 
+
+if not os.path.isfile(debugFileName):
+    with open(debugFileName, "w") as f:
+        f.write("\n")
+
 def colored(msg, label="INFO") :
     """
     Return a colored string. Formatting is optional.
@@ -38,6 +49,13 @@ def colored(msg, label="INFO") :
     return "{0} {1}".format(color+msg, ENDC)
 
 def printDebug(label, msg, msg2='', frame=None, exception=None):
+    """Dump the content into a logfile or to stderr """
+
+    if label == "LOG":
+        with open(debugFileName, "a") as logF:
+            logF.write(msg+'\n')
+        return 
+
     if not frame :
         print("[%s] %s %s" % (label, colored(msg,label),msg2)
                 , file=sys.stderr)
@@ -51,6 +69,7 @@ def printDebug(label, msg, msg2='', frame=None, exception=None):
                                              )
               , file=sys.stderr
               )
+
     if exception:
         print(" [Expcetion] {0}".format(exception))