+++ CHANGES.rst
@@ -11,3 +11,8 @@ Wed 29 Jan 2014 05:53:58 PM IST
 - The YAML front matter in post is closely match with jekyll. Same post can be
   uploaded to github page.
 
+Sat 01 Feb 2014 12:10:25 AM IST
+
+- The defualt config file is ``~/.config/twordpress/config``
+- We can now post latex on wordpress. Use $latex_formulae$ to write latex in
+  markdown.